Created utils for unittest package, move some helper functions
