+++ cmd/util/cmd/epochs/cmd/reset_test.go
@@ -14,7 +14,6 @@ import (
 	"github.com/onflow/flow-go/model/bootstrap"
 	"github.com/onflow/flow-go/state/protocol/inmem"
 	"github.com/onflow/flow-go/utils/unittest"
-	"github.com/onflow/flow-go/utils/unittest/generator"
 )
 
 // TestReset_LocalSnapshot tests the command with a local snapshot file.
@@ -48,7 +47,7 @@ func TestReset_LocalSnapshot(t *testing.T) {
 
 			// compare to expected values
 			expectedArgs := extractResetEpochArgs(rootSnapshot)
-			generator.VerifyCdcArguments(t, expectedArgs, outputTxArgs)
+			unittest.VerifyCdcArguments(t, expectedArgs, outputTxArgs)
 		})
 	})
 
@@ -96,7 +95,7 @@ func TestReset_BucketSnapshot(t *testing.T) {
 		rootSnapshot, err := getSnapshotFromBucket(fmt.Sprintf(rootSnapshotBucketURL, flagBucketNetworkName))
 		require.NoError(t, err)
 		expectedArgs := extractResetEpochArgs(rootSnapshot)
-		generator.VerifyCdcArguments(t, expectedArgs, outputTxArgs)
+		unittest.VerifyCdcArguments(t, expectedArgs, outputTxArgs)
 	})
 
 	// should output arguments to stdout, including specified payout
@@ -118,7 +117,7 @@ func TestReset_BucketSnapshot(t *testing.T) {
 		rootSnapshot, err := getSnapshotFromBucket(fmt.Sprintf(rootSnapshotBucketURL, flagBucketNetworkName))
 		require.NoError(t, err)
 		expectedArgs := extractResetEpochArgs(rootSnapshot)
-		generator.VerifyCdcArguments(t, expectedArgs, outputTxArgs)
+		unittest.VerifyCdcArguments(t, expectedArgs, outputTxArgs)
 	})
 
 	// with a missing snapshot, should log an error
